export-worker/requirements-dev.txt
# Development & Testing Dependencies
# Install: pip install -r requirements-dev.txt

-r requirements.txt

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Redis mock
fakeredis[aioredis]==2.21.1

# Code quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
